:truck: Move option values to own module
